+++ packages/api/src/utils/router.js
@@ -1,5 +1,5 @@
 import { parse } from 'regexparam'
-import { database } from '../constants.js'
+import { database, GATEWAY_NFT_STORAGE_HOSTNAME } from '../constants.js'
 
 /**
  * @typedef {{ params: Record<string, string> }} BasicRouteContext
@@ -184,8 +184,11 @@ class Router {
     // Add more if needed for other backends
     const passThrough = [database.url]
 
-    // Ignore http requests from the passthrough list above
-    if (!passThrough.includes(`${url.protocol}//${url.host}`)) {
+    // Ignore http requests from the passthrough list above and gateway
+    if (
+      !passThrough.includes(`${url.protocol}//${url.host}`) &&
+      !url.hostname.includes(GATEWAY_NFT_STORAGE_HOSTNAME)
+    ) {
       event.respondWith(this.route(event))
     }
   }