+++ packages/gateway/src/metrics.js
@@ -5,9 +5,10 @@ import pMap from 'p-map'
 import {
   METRICS_CACHE_MAX_AGE,
   SUMMARY_METRICS_ID,
+  REDIRECT_COUNTER_METRICS_ID,
   HTTP_STATUS_SUCCESS,
 } from './constants.js'
-import { histogram } from './durable-objects/gateway-metrics.js'
+import { responseTimeHistogram } from './utils/histogram.js'
 import { contentLengthHistogram } from './durable-objects/summary-metrics.js'
 
 /**
@@ -17,6 +18,7 @@ import { contentLengthHistogram } from './durable-objects/summary-metrics.js'
  * @typedef MetricsDurable
  * @property {SummaryMetrics} summaryMetrics
  * @property {Record<string,GatewayMetrics>} ipfsGateways
+ * @property {number} gatewayRedirectCount
  */
 
 /**
@@ -28,40 +30,50 @@ import { contentLengthHistogram } from './durable-objects/summary-metrics.js'
  * @returns {Promise<Response>}
  */
 export async function metricsGet(request, env, ctx) {
-  // TODO: Set cache
-  // const cache = caches.default
-  // let res = await cache.match(request)
+  const cache = caches.default
+  let res = await cache.match(request)
 
-  // if (res) {
-  //   return res
-  // }
-  let res
+  if (res) {
+    return res
+  }
 
-  const [summaryMetrics, ipfsGateways] = await Promise.all([
-    (async () => {
-      const id = env.summaryMetricsDurable.idFromName(SUMMARY_METRICS_ID)
-      const stub = env.summaryMetricsDurable.get(id)
+  const [summaryMetrics, ipfsGateways, gatewayRedirectCount] =
+    await Promise.all([
+      (async () => {
+        const id = env.summaryMetricsDurable.idFromName(SUMMARY_METRICS_ID)
+        const stub = env.summaryMetricsDurable.get(id)
 
-      const stubResponse = await stub.fetch(request)
-      /** @type {SummaryMetrics} */
-      const summaryMetrics = await stubResponse.json()
+        const stubResponse = await stub.fetch(request)
+        /** @type {SummaryMetrics} */
+        const summaryMetrics = await stubResponse.json()
 
-      return summaryMetrics
-    })(),
-    pMap(env.ipfsGateways, async (gw) => {
-      const id = env.gatewayMetricsDurable.idFromName(gw)
-      const stub = env.gatewayMetricsDurable.get(id)
+        return summaryMetrics
+      })(),
+      pMap(env.ipfsGateways, async (gw) => {
+        const id = env.gatewayMetricsDurable.idFromName(gw)
+        const stub = env.gatewayMetricsDurable.get(id)
 
-      const stubResponse = await stub.fetch(request)
-      /** @type {GatewayMetrics} */
-      const gwMetrics = await stubResponse.json()
+        const stubResponse = await stub.fetch(request)
+        /** @type {GatewayMetrics} */
+        const gwMetrics = await stubResponse.json()
 
-      return {
-        gwMetrics,
-        gw,
-      }
-    }),
-  ])
+        return {
+          gwMetrics,
+          gw,
+        }
+      }),
+      (async () => {
+        const id = env.gatewayRedirectCounter.idFromName(
+          REDIRECT_COUNTER_METRICS_ID
+        )
+        const stub = env.gatewayRedirectCounter.get(id)
+
+        const stubResponse = await stub.fetch(request)
+        const { gatewayRedirectCount } = await stubResponse.json()
+
+        return gatewayRedirectCount
+      })(),
+    ])
 
   /** @type {MetricsDurable} */
   const metricsCollected = {
@@ -70,6 +82,7 @@ export async function metricsGet(request, env, ctx) {
       (obj, item) => Object.assign(obj, { [item.gw]: item.gwMetrics }),
       {}
     ),
+    gatewayRedirectCount,
   }
 
   const totalResponsesPerGateway = []
@@ -80,9 +93,26 @@ export async function metricsGet(request, env, ctx) {
   })
 
   const metrics = [
+    `# HELP nftgateway_summary_responses_total Total winner and cached responses returned.`,
+    `# TYPE nftgateway_summary_responses_total counter`,
+    `nftgateway_summary_responses_total{env="${env.ENV}"} ${
+      metricsCollected.summaryMetrics.totalWinnerSuccessfulRequests +
+      metricsCollected.summaryMetrics.totalCachedResponses
+    }`,
+    `# HELP nftgateway_summary_response_time_seconds_total Accumulated summary response time of winner responses and cached responses.`,
+    `# TYPE nftgateway_summary_response_time_seconds_total summary`,
+    `nftgateway_summary_response_time_seconds_total{env="${env.ENV}"} ${msToS(
+      metricsCollected.summaryMetrics.totalWinnerResponseTime +
+        metricsCollected.summaryMetrics.totalCachedResponseTime
+    )}`,
     `# HELP nftgateway_cache_hit_responses_total Total cached responses returned.`,
     `# TYPE nftgateway_cache_hit_responses_total counter`,
     `nftgateway_cache_hit_responses_total{env="${env.ENV}"} ${metricsCollected.summaryMetrics.totalCachedResponses}`,
+    `# HELP nftgateway_cache_hit_response_time_seconds_total Accumulated cached response time.`,
+    `# TYPE nftgateway_cache_hit_response_time_seconds_total summary`,
+    `nftgateway_cache_hit_response_time_seconds_total{env="${env.ENV}"} ${msToS(
+      metricsCollected.summaryMetrics.totalCachedResponseTime
+    )}`,
     `# HELP nftgateway_winner_requests_total Total winner requests.`,
     `# TYPE nftgateway_winner_requests_total counter`,
     `nftgateway_winner_requests_total{env="${env.ENV}"} ${metricsCollected.summaryMetrics.totalWinnerSuccessfulRequests}`,
@@ -91,6 +121,18 @@ export async function metricsGet(request, env, ctx) {
     `nftgateway_winner_response_time_seconds_total{env="${env.ENV}"} ${msToS(
       metricsCollected.summaryMetrics.totalWinnerResponseTime
     )}`,
+    `# HELP nftgateway_summary_responses_per_time_total total of responses per response time bucket`,
+    `# TYPE nftgateway_summary_responses_per_time_total histogram`,
+    ...responseTimeHistogram.map(
+      (t) =>
+        `nftgateway_summary_responses_per_time_total{le="${msToS(t)}",env="${
+          env.ENV
+        }"} ${metricsCollected.summaryMetrics.responseTimeHistogram[t]}`
+    ),
+    `nftgateway_summary_responses_per_time_total{le="+Inf",env="${env.ENV}"} ${
+      metricsCollected.summaryMetrics.totalWinnerSuccessfulRequests +
+      metricsCollected.summaryMetrics.totalCachedResponses
+    }`,
     `# HELP nftgateway_response_time_seconds_total Accumulated response time of each gateway.`,
     `# TYPE nftgateway_response_time_seconds_total summary`,
     ...env.ipfsGateways.map(
@@ -184,11 +226,15 @@ export async function metricsGet(request, env, ctx) {
     ),
     `# HELP nftgateway_requests_per_time_total total of requests per response time bucket`,
     `# TYPE nftgateway_requests_per_time_total histogram`,
-    ...histogram.map((t) => {
+    ...responseTimeHistogram.map((t) => {
       return env.ipfsGateways
         .map(
           (gw) =>
-            `nftgateway_requests_per_time_total{gateway="${gw}",le="${t}",env="${env.ENV}"} ${metricsCollected.ipfsGateways[gw].responseTimeHistogram[t]}`
+            `nftgateway_requests_per_time_total{gateway="${gw}",le="${msToS(
+              t
+            )}",env="${env.ENV}"} ${
+              metricsCollected.ipfsGateways[gw].responseTimeHistogram[t]
+            }`
         )
         .join('\n')
     }),
@@ -208,13 +254,19 @@ export async function metricsGet(request, env, ctx) {
       (t) =>
         `nftgateway_responses_content_length_total{le="${t}",env="${env.ENV}"} ${metricsCollected.summaryMetrics.contentLengthHistogram[t]}`
     ),
-    `nftgateway_responses_content_length_total{le="+Inf",env="${env.ENV}"} ${metricsCollected.summaryMetrics.totalWinnerSuccessfulRequests}`,
+    `nftgateway_responses_content_length_total{le="+Inf",env="${env.ENV}"} ${
+      metricsCollected.summaryMetrics.totalWinnerSuccessfulRequests +
+      metricsCollected.summaryMetrics.totalCachedResponses
+    }`,
     `# HELP nftgateway_responses_content_length_bytes_total Accumulated content length of delivered cached responses`,
     `# TYPE nftgateway_responses_content_length_bytes_total summary`,
     `nftgateway_responses_content_length_bytes_total{env="${env.ENV}"} ${metricsCollected.summaryMetrics.totalContentLengthBytes}`,
     `# HELP nftgateway_cached_responses_content_length_bytes_total Accumulated content length of delivered cached responses`,
     `# TYPE nftgateway_cached_responses_content_length_bytes_total summary`,
     `nftgateway_cached_responses_content_length_bytes_total{env="${env.ENV}"} ${metricsCollected.summaryMetrics.totalCachedContentLengthBytes}`,
+    `# HELP nftgateway_redirect_total Total redirects to gateway.`,
+    `# TYPE nftgateway_redirect_total counter`,
+    `nftgateway_redirect_total{env="${env.ENV}"} ${metricsCollected.gatewayRedirectCount}`,
   ].join('\n')
 
   res = new Response(metrics, {
@@ -223,7 +275,7 @@ export async function metricsGet(request, env, ctx) {
     },
   })
 
-  // ctx.waitUntil(cache.put(request, res.clone()))
+  ctx.waitUntil(cache.put(request, res.clone()))
 
   return res
 }