+++ packages/website/pages/login.js
@@ -1,5 +1,5 @@
 import { useState } from 'react'
-import Router, { useRouter } from 'next/router'
+import Router from 'next/router'
 import { loginEmail, loginSocial } from '../lib/magic.js'
 import countly from '../lib/countly'
 import Button from '../components/button.js'
@@ -21,8 +21,6 @@ export default function Login() {
   const [errorMsg, setErrorMsg] = useState('')
   const [disabled, setDisabled] = useState(false)
   const [isRedirecting, setIsRedirecting] = useState(false)
-  const { query } = useRouter()
-  const version = /** @type {string} */ (query.version)
 
   /**
    * @param {import('react').ChangeEvent<HTMLFormElement>} e
@@ -34,7 +32,7 @@ export default function Login() {
     setDisabled(true)
 
     try {
-      await loginEmail(e.currentTarget.email.value, version)
+      await loginEmail(e.currentTarget.email.value)
       await queryClient.invalidateQueries('magic-user')
       Router.push('/files')
     } catch (/** @type {any} */ error) {
@@ -86,7 +84,7 @@ export default function Login() {
             className="w5"
             onClick={() => {
               setIsRedirecting(true)
-              loginSocial('github', version)
+              loginSocial('github')
             }}
             tracking={{
               event: countly.events.LOGIN_CLICK,